added check to see if logfile dataURL is null

+++ datawolf-core/src/main/java/edu/illinois/ncsa/datawolf/Executor.java
@@ -285,7 +285,9 @@ public abstract class Executor {
                 logFileDao.save(logfile);
             } else {
                 // Delete existing file
-                fileStorage.deleteFile(logfile.getLog());
+                if (logfile.getLog().getDataURL() != null) {
+                    fileStorage.deleteFile(logfile.getLog());
+                }
             }
 
             try {
